AOC-816 ITUATB-64 Local item table should only show the items in their specific sites. i.e. Vietnam should only show Vietnam items but not south china and Suzhou.

+++ src/com/avery/storage/dao/impl/OrgDaoImpl.java
@@ -81,7 +81,7 @@ public class OrgDaoImpl extends GenericDaoImpl<Org, Long> implements
 		if (queryMap.getFirst("siteId") != null && queryMap.getFirst("siteId") != "") {
 			String siteId = (String) queryMap.getFirst("siteId");
 			if (!siteId.equals("") && siteId != null && !siteId.isEmpty())
-				criteria.add(Restrictions.eq("site.id", Integer.parseInt(siteId)));
+				criteria.add(Restrictions.eq("site.id", Long.parseLong(siteId)));
 		}
 		entitiesMap.put("org", criteria.list());
 		return entitiesMap;
